requirements.txt
opencv-python
numpy<2.0
pandas
tf-keras
deepface
customtkinter
pillow
streamlit
watchdogas
